device/requirements.txt
gpsd-py3==0.3.0
pynmea2==1.18.0
requests==2.28.1
schedule==1.1.0 